requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
google-api-python-client==2.114.0
google-auth==2.27.0
python-dotenv==1.0.0
pyyaml==6.0.1
email-validator==2.1.0
cachetools==5.3.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
